+++ CHANGELOG.md
@@ -65,6 +65,8 @@
   - Type `NodeInfo` field `details` is now wrapped in `Upward`.
   - Type `Peer` field `consensus_info` is now wrapped in `Upward`.
   - Type `PeerConsensusInfo::Node` unnamed field is now wrapped in `Upward`.
+  - Type `AccountInfo` field `account_stake` changes from `Option<AccountStakingInfo>` to `Option<Upward<AccountStakingInfo>>`.
+  - Type `Cooldown` field `status` is now wrapped in `Upward`.
 
 ## 7.0.0